pyproject.toml
[project]
name = "locust-user"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "locust>=2.41.6",
    "paho-mqtt>=2.1.0",
]
